MBS-7043: Keep DiscID when it is moved onto itself
The code to apply a Move DiscID edit checked whether the DiscID was
already attached to the target medium, and if so, simply removed it
from the source medium. If source and target medium were identical,
this meant that the DiscID was lost.

This is fixed by only removing the DiscID if source and target are
different.

+++ lib/MusicBrainz/Server/Edit/Medium/MoveDiscID.pm
@@ -158,9 +158,10 @@ sub accept
         )
     }
 
-    if ($self->c->model('MediumCDTOC')->medium_has_cdtoc(
-        $medium->id,
-        $medium_cdtoc->cdtoc
+    if ($self->data->{old_medium}{id} != $self->data->{new_medium}{id} &&
+        $self->c->model('MediumCDTOC')->medium_has_cdtoc(
+            $medium->id,
+            $medium_cdtoc->cdtoc
     )) {
         $self->c->model('MediumCDTOC')->delete(
             $self->data->{medium_cdtoc}{id}
